+++ src/components/spartan/SPageTitle/index.ts
@@ -1,0 +1,3 @@
+export { default as SPageTitle } from './SPageTitle.vue';
+
+export type { TPageTitleProps } from './types';