+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2.0.0-beta.67 - 2024-02-12
 ### Added
-- New `SMethodIcon` component.
+- New `SPageTitle` component.
+- New `SCardBrand` component.
 - `SSectionDescription` now supports class merging.
 
 ## 2.0.0-beta.66 - 2024-02-12